ansible/requirements.txt
ansible>=7.0.0
molecule>=5.0.0
molecule-docker
pytest-testinfra
ansible-lint
yamllint
flake8
pytest-molecule
